+++ js/form-img.js
@@ -51,85 +51,3 @@ userModalCloseElement.addEventListener('keydown', (evt) => {
     closeUserModal();
   }
 });
-
-
-//Изменение размера изображения
-
-function removeLastChar (str) {
-  return str ? str.slice(0 , -1) : str;
-}
-
-function toNumber (str) {
-  return Number(str);
-}
-
-const scaleControlSmaller = document.querySelector('.scale__control--smaller');
-const scaleControlBigger = document.querySelector('.scale__control--bigger');
-const imgPreview = document.querySelector('.img-upload__preview');
-const input = document.querySelector('input[name="scale"]');
-const maxScale = 100;
-const minScale = 25;
-const step = 25;
-
-scaleControlBigger.addEventListener('click', () => {
-  const currentValue = input.value;
-  const numberValue = toNumber(removeLastChar(currentValue));
-
-  if (numberValue && numberValue < maxScale) {
-    const newValue = numberValue + step;
-
-    input.value = `${newValue}%`;
-    imgPreview.style.transform = `scale(${newValue / 100})`;
-  }
-});
-
-scaleControlSmaller.addEventListener('click', () => {
-  const currentValue = input.value;
-  const numberValue = toNumber(removeLastChar(currentValue));
-
-  if (numberValue && numberValue > minScale) {
-    const newValue = numberValue - step;
-
-    input.value = `${newValue}%`;
-    imgPreview.style.transform = `scale(${newValue / 100})`;
-  }
-});
-
-//Изменение фильтров на фото
-
-
-const effectsListElement = document.querySelector('.effects__list');
-const radioEffects = document.querySelectorAll('input[name="effect"]');
-
-effectsListElement.addEventListener('click', (index) => {
-  const { target } = index;
-  if (imgPreview.classList.length > 1) {
-    const currentEffect = imgPreview.classList[1];
-
-    imgPreview.classList.remove(currentEffect);
-  }
-
-  imgPreview.classList.add(`effects__preview--${target?.value}`);
-  radioEffects.value = target?.value;
-});
-
-
-/* global Pristine:readonly */
-const photoForm = document.querySelector('.img-upload__form');
-
-const pristine = new Pristine(photoForm,{
-  classTo: 'img-upload__text text',
-  errorTextParent: 'img-upload__text text',
-  errorTextClass: 'img-upload__text text',
-});
-
-photoForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-
-  const isValid = pristine.validate();
-  if (isValid) {
-    console.log('Можно отправлять');
-  } else {
-    console.log('Форма невалидна');
-  }
-});
