+++ js/main.js
@@ -2,3 +2,6 @@ import './picture.js';
 import './util.js';
 import './data.js';
 import './form-img.js';
+import './size-photo.js';
+import './validity.js';
+import './photo-filters.js';
